Feat: Header TypeScript 적용 및 커스텀 이벤트 변경
